reworked the plugin hierarchy to make room for the recorder. Also remove duplicated code between tracker plugin and player plugin.
