+++ lib/src/plugins/sound_recorder_plugin.dart
@@ -18,7 +18,7 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 
-import '../impl/sound_recorder_impl.dart';
+import '../sound_recorder.dart';
 import 'base_plugin.dart';
 
 /// Provides communications with the platform
@@ -35,7 +35,7 @@ class SoundRecorderPlugin extends BasePlugin {
       : super('com.dooboolab.flutter_sound_recorder');
 
   Future<dynamic> onMethodCallback(
-      covariant SoundRecorderImpl connector, MethodCall call) {
+      covariant SoundRecorder connector, MethodCall call) {
     switch (call.method) {
       case "updateRecorderProgress":
         {
